# openapi.yaml
openapi: 3.0.3
info:
  title: Airport Transfer Marketplace API
  version: 1.1.0
  description: |
    Independent Airport Transfer Marketplace API.
    Covers authentication, suppliers, airports/zones/routes, pricing, bookings, rides, and payments.

    ## Role-Based Access Control

    - `GET /public/*`, `POST /public/*`, `/webhooks/*`: **no authentication**
    - `/auth/*`: no authentication
    - `/supplier/*`: requires JWT with roles: `SUPPLIER_OWNER`, `DISPATCHER`, or `DRIVER`
    - `/admin/*`: requires JWT with role: `ADMIN`

    ## Currency Handling

    - Marketplace operates in **EUR** as base currency
    - Suppliers can set prices in their local currency (TRY, USD, GBP, etc.)
    - Currency conversion happens at booking time using real-time rates
    - Supplier payouts are made in their configured payout currency

    ## Cancellation Policy

    - Free cancellation up to **24 hours** before pickup
    - **50% refund** for cancellations **12–24 hours** before pickup
    - **No refund** for cancellations less than **12 hours** before pickup
    - Supplier cancellations result in full refund + possible compensation
servers:
  - url: https://api.airporttransferportal.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RouteDirection:
      type: string
      enum: [FROM_AIRPORT, TO_AIRPORT, BOTH]
...
    Airport:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
          example: IST
        name:
          type: string
        city:
          type: string
        country:
          type: string
        timezone:
          type: string
          example: Europe/Istanbul
        isActive:
          type: boolean
    Zone:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Sultanahmet
        city:
          type: string
        country:
          type: string
        isActive:
          type: boolean
    Route:
      type: object
      properties:
        id:
          type: integer
          format: int64
        airport:
          $ref: '#/components/schemas/Airport'
        zone:
          $ref: '#/components/schemas/Zone'
        direction:
          $ref: '#/components/schemas/RouteDirection'
        approxDistanceKm:
          type: number
          format: double
        approxDurationMin:
          type: integer
        isActive:
          type: boolean
    Supplier:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        country:
          type: string
        city:
          type: string
        taxNumber:
          type: string
        legalName:
          type: string
        isVerified:
          type: boolean
        commissionRate:
          type: number
          format: double
...
    RouteCreate:
      type: object
      required: [airportId, zoneId, direction]
      properties:
        airportId:
          type: integer
          format: int64
        zoneId:
          type: integer
          format: int64
        direction:
          $ref: '#/components/schemas/RouteDirection'
        approxDistanceKm:
          type: number
          format: double
        approxDurationMin:
          type: integer
    SupplierCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        country:
          type: string
        city:
          type: string
        taxNumber:
          type: string
        legalName:
          type: string
        commissionRate:
          type: number
          format: double
...
    FlightTrackingStatus:
      type: string
      enum: [SCHEDULED, DEPARTED, EN_ROUTE, LANDED, ARRIVED_GATE, CANCELLED, DIVERTED, DELAYED, UNKNOWN]
...
    FlightTracking:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        bookingCode:
          type: string
        flightNumber:
          type: string
        flightDate:
          type: string
          format: date
        departureAirport:
          type: string
        arrivalAirport:
          type: string
        scheduledDeparture:
          type: string
          format: date-time
          nullable: true
        scheduledArrival:
          type: string
          format: date-time
        estimatedArrival:
          type: string
          format: date-time
          nullable: true
        actualArrival:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/FlightTrackingStatus'
        delayMinutes:
          type: integer
        delayReason:
          type: string
          nullable: true
        terminal:
          type: string
          nullable: true
        gate:
          type: string
          nullable: true
        baggageClaim:
          type: string
          nullable: true
        trackingSource:
          type: string
          nullable: true
        lastChecked:
          type: string
          format: date-time
          nullable: true
        customerName:
          type: string
        pickupTime:
          type: string
          format: date-time
        pickupAdjusted:
          type: boolean
...
paths:
...
  /admin/routes:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      x-roles:
        - ADMIN
      summary: List all routes
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Route'
    post:
      tags: [Admin]
      security:
        - bearerAuth: []
      x-roles:
        - ADMIN
      summary: Create route between airport and zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteCreate'
      responses:
        '201':
          description: Route created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Route'
  /admin/suppliers:
    get:
      tags: [Admin]
      security:
        - bearerAuth: []
      x-roles:
        - ADMIN
      summary: List all suppliers
      responses:
        '200':
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'
    post:
      tags: [Admin]
      security:
        - bearerAuth: []
      x-roles:
        - ADMIN
      summary: Create supplier company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '201':
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
...
  /dispatch/flights:
    get:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Get flight tracking data
      description: Returns flight status for upcoming bookings
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, delayed, arriving, landed]
          description: Filter flights by status
      responses:
        '200':
          description: Flight tracking data
          content:
            application/json:
              schema:
                type: object
                properties:
                  flights:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlightTracking'
    post:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Add flight tracking
      description: Add flight tracking for a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, flightNumber, flightDate]
              properties:
                bookingId:
                  type: integer
                flightNumber:
                  type: string
                flightDate:
                  type: string
                  format: date
                departureAirport:
                  type: string
                arrivalAirport:
                  type: string
                scheduledDeparture:
                  type: string
                  format: date-time
                scheduledArrival:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Flight tracking added
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
    put:
      tags: [Dispatch]
      security:
        - bearerAuth: []
      summary: Update flight tracking
      description: Update flight status and arrival info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id:
                  type: integer
                status:
                  $ref: '#/components/schemas/FlightTrackingStatus'
                estimatedArrival:
                  type: string
                  format: date-time
                actualArrival:
                  type: string
                  format: date-time
                delayMinutes:
                  type: integer
                delayReason:
                  type: string
                terminal:
                  type: string
                gate:
                  type: string
                baggageClaim:
                  type: string
                trackingSource:
                  type: string
      responses:
        '200':
          description: Flight tracking updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...